IT HAPPENED AGAIN NOO
redid course to use updated models, as well as make it easier to read.
moved login to its own dedicated file and removed admin.
made semester easier to understand by changing the formatting.
redid staff to make the formatting easier as well as updating it to use the new models.